+++ src/time_split/support/__init__.py
@@ -13,11 +13,9 @@ dependencies if you need to use the ``support`` module.
 
 from .._backend import DatetimeIndexSplitter, expand_limits, process_available
 from .._frontend import default_metrics_formatter, fold_weight, format_expanded_limits, to_string
-from ._create_explorer_link import create_explorer_link
 
 __all__ = [
     "DatetimeIndexSplitter",
-    "create_explorer_link",
     "default_metrics_formatter",
     "expand_limits",
     "fold_weight",